pyproject.toml
[tool.poetry]
name = "nlp-multiagent-rag"
version = "0.1.0"
description = ""
authors = ["Rino Albertin <rino.albertin@ost.ch>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
unstructured = { extras = ["pdf"], version = "^0.18.2" }
python-dotenv = "^1.1.1"
pdfplumber = "^0.11.7"
tiktoken = "^0.9.0"
openai = "^1.93.0"
faiss-cpu = "^1.11.0"
streamlit = "^1.46.1"
sentence-transformers = "^5.0.0"
redis = "^6.2.0"
langgraph = "^0.5.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.12.1"
pre-commit = "^4.2.0"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
